fix(signals): browser-safe entry so the web bundle excludes server-only use-cases
The web (Vite browser condition) bundles @domain/signals, and createSignalUseCase
transitively pulls the judge template -> @domain/ai, which isn't browser-safe
(createEvaluationUseCase isn't in @domain/evaluations' browser entry). Split the
package like @domain/evaluations: browser.ts holds the browser-safe surface (types,
ports, pure helpers, read use-cases) and index.ts re-exports it plus the four
server-only mutation use-cases (create/update/delete/backfill). The 'browser' export
condition points at browser.ts; API/workers/tests keep using the full index.

Phase 2 / PR3.
