+++ packages/domain/signals/package.json
@@ -6,7 +6,10 @@
   "main": "src/index.ts",
   "types": "src/index.ts",
   "exports": {
-    ".": "./src/index.ts",
+    ".": {
+      "browser": "./src/browser.ts",
+      "default": "./src/index.ts"
+    },
     "./testing": "./src/testing/index.ts"
   },
   "scripts": {
